move merkleization into ssz, add and fix more multi-proof code from spec
